Rename: api, interface, url 파일 별도 분리
