+++ js/interface/postInterface.ts
